:recycle: Move ResultCompetence to his domain
