+++ api/tests/certification/results/unit/domain/models/ResultCompetenceTree_test.js
@@ -1,5 +1,5 @@
+import { ResultCompetence } from '../../../../../../src/certification/results/domain/models/ResultCompetence.js';
 import { ResultCompetenceTree } from '../../../../../../src/certification/results/domain/models/ResultCompetenceTree.js';
-import { ResultCompetence } from '../../../../../../src/shared/domain/models/ResultCompetence.js';
 import { domainBuilder, expect } from '../../../../../test-helper.js';
 
 describe('Unit | Domain | Models | ResultCompetence', function () {
