+++ api/tests/certification/results/unit/infrastructure/serializers/shareable-certificate-serializer_test.js
@@ -1,6 +1,6 @@
+import { ResultCompetence } from '../../../../../../src/certification/results/domain/models/ResultCompetence.js';
 import { ResultCompetenceTree } from '../../../../../../src/certification/results/domain/models/ResultCompetenceTree.js';
 import * as serializer from '../../../../../../src/certification/results/infrastructure/serializers/shareable-certificate-serializer.js';
-import { ResultCompetence } from '../../../../../../src/shared/domain/models/index.js';
 import { domainBuilder, expect } from '../../../../../test-helper.js';
 
 describe('Unit | Serializer | JSONAPI | shareable-certificate-serializer', function () {
